fix tests

+++ hilland_backend/controllers/news.js
@@ -18,6 +18,15 @@ newsRouter.get('/:id', async(request, response) => {
 })
   
 newsRouter.put('/:id',  (request, response, next) => {
+  const token = request.token 
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
+  if (decodedToken.role !== 'admin'){
+    return response.status(401).json({ error: 'you don´t have rights for this operation' })
+  } 
   const news = ({
     title: request.body.title,
     content: request.body.content,
@@ -33,10 +42,13 @@ newsRouter.put('/:id',  (request, response, next) => {
 })
   
 newsRouter.post('/', async (request, response) => {
-  const token = request.token //getTokenFrom(request)
+  const token = request.token 
   const decodedToken = jwt.verify(token, process.env.SECRET)
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  if (request.user.role !== 'admin'){
+    return response.status(401).json({ error: 'you don´t have rights for this operation' })
   }
   
   const news = new News({
@@ -52,6 +64,15 @@ newsRouter.post('/', async (request, response) => {
 })
 
 newsRouter.delete('/:id', async (request, response) => {
+  const token = request.token 
+  
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+ /*  if (request.user.role !== 'admin'){
+    return response.status(401).json({ error: 'you don´t have rights for this operation' })
+  } */
   await News.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
